arm64: Add explicit cast to arm64_ftr_value
arm64_ftr_value() takes an unsigned int and returns it as s64. Add
an explicit conversion to allow building with -Wextra error level.

Reviewed-on: http://git-master/r/1251408
GVS: Gerrit_Virtual_Submit

+++ arch/arm64/include/asm/cpufeature.h
@@ -149,7 +149,7 @@ static inline s64 arm64_ftr_value(struct arm64_ftr_bits *ftrp, u64 val)
 {
 	return ftrp->sign ?
 		cpuid_feature_extract_field_width(val, ftrp->shift, ftrp->width) :
-		cpuid_feature_extract_unsigned_field_width(val, ftrp->shift, ftrp->width);
+		(s64)cpuid_feature_extract_unsigned_field_width(val, ftrp->shift, ftrp->width);
 }
 
 static inline bool id_aa64mmfr0_mixed_endian_el0(u64 mmfr0)
